+++ mcRBM_matlab/train_mcRBM.m
@@ -4,7 +4,7 @@ function [W,VF,FH,vb,hb_cov,hb_mean,hmc_step, hmc_ave_rej] = train_mcRBM(X,W,VF,
 
     small = 0.5;
     normVF = 1;
-    negdata = randn(num_vis,batch_size,'double');
+    negdata = randn(num_vis,batch_size,'single');
     meanEnergy = zeros(num_epochs,1);
     minEnergy = zeros(num_epochs,1);
     maxEnergy = zeros(num_epochs,1);
